requirements.txt
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dateutil==2.8.2
numpy==1.26.2
pyarrow==14.0.2
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
tabulate==0.9.0
PyGithub==1.59.1
python-dotenv==1.0.0

